% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flate_gdp.R
\name{flate_gdp}
\alias{flate_gdp}
\title{Inflate or Deflate by GDP}
\usage{
flate_gdp(price, from, to)
}
\arguments{
\item{price}{in pounds (GBP, numeric)}

\item{from}{financial year of the price (numeric, `2018` is the 2018/19
financial year)}

\item{to}{financial year to inflate to (integer, `2019` is the 2019/20
financial year)}
}
\description{
Given a UK price in GBP, inflates to from the given year to the given year
(in the future), or deflates from the given year to the given year (in the
past).
}
\examples{
x <- flate_gdp(100, from = 2017, to = 2019)
x
flate_gdp(x, from = 2019, to = 2017)
}
