% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdp_deflator}
\alias{gdp_deflator}
\title{GDP Deflator}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 68 rows and 2 columns.}
\source{
HM Treasury: GDP deflators at market prices, and money GDP June 2018
(Quarterly National Accounts)
<https://www.gov.uk/government/statistics/gdp-deflators-at-market-prices-and-money-gdp-june-2018-quarterly-national-accounts>

The source is distributed with this package.  Find its location on your
system as follows:

```
system.file("extdata", "gdp-deflator.csv", package = "electionfunding")
```
}
\usage{
gdp_deflator
}
\description{
Deflators per financial year, from HM Treasury based on national statistics
by the Office for National Statistics.
}
\keyword{datasets}
